modify formatto print

+++ experiments/bytewise_conjunction.cpp
@@ -29,7 +29,7 @@ int main(){
 	for(size_t i = 0; i < num_rows; i++){
         ByteUnit code = std::rand() & mask;
         column1->SetTuple(i, code);
-        std::cout << code << " " << std::endl;
+        std::cout << static_cast<uint16_t>(code) << " " << std::endl;
     }
 
 	BytewiseScan scan;
